webserver/src/requirements.txt
flask==3.0.1
Werkzeug==3.0.1
mysql-connector-python==8.1.0
flake8==7.0.0
requests==2.26.0
docker-py==1.8
